:recycle: change business logic from user to auth
