+++ src/app/auth/login/login.component.ts
@@ -3,7 +3,7 @@ import { FormBuilder, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
 
-import { UserService } from 'src/app/core/services/user.service';
+import { AuthService } from '@services/auth.service';
 
 @Component({
   selector: 'app-login',
@@ -20,7 +20,7 @@ export class LoginComponent implements OnInit {
   constructor(
     private fb: FormBuilder,
     private router: Router,
-    private userService: UserService,
+    private authService: AuthService,
   ) { }
 
   ngOnInit(): void {
@@ -28,13 +28,11 @@ export class LoginComponent implements OnInit {
 
   login() {
     if(this.loginForm.valid) {
-      this.userService.login(this.loginForm.value).subscribe({
+      this.authService.login(this.loginForm.value).subscribe({
         next: () => {
-          console.log('entro en next');
           this.router.navigate(['/']);
         },
         error: () => {
-          console.log('se fue a swal');
           Swal.fire('Usuario y/o password incorrectos', '', 'error');
         }
       })
